+++ src/Bpp/Seq/AlphabetIndex/AAChouFasmanAHelixIndex.h
@@ -92,12 +92,12 @@ public:
   double getIndex(int state) const throw (BadIntException)
   {
     if (state < 0 || state > 19) throw BadIntException(state, "AAChouFasmanAHelixIndex::getIndex(). Invalid state.", &AlphabetTools::PROTEIN_ALPHABET);
-    return aHelix_[state];
+    return aHelix_[static_cast<size_t>(state)];
   }
 
   double getIndex(const std::string& state) const throw (BadCharException)
   {
-    return aHelix_[AlphabetTools::PROTEIN_ALPHABET.charToInt(state)];
+    return aHelix_[static_cast<size_t>(AlphabetTools::PROTEIN_ALPHABET.charToInt(state))];
   }
 
   std::vector<double>* getIndexVector() const { return new std::vector<double>(aHelix_); }
